requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiogram2==0.1.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
certifi==2025.1.31
defusedxml==0.7.1
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
numpy==2.2.4
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxlsb==1.0.10
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.18.3
